backend/requirements.txt
# backend/requirements.txt - Korrigierte Dependencies

# === FASTAPI & CORE ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# === DATABASE & EMBEDDINGS ===
chromadb==0.4.18
sentence-transformers==2.2.2
# Alternative: transformers==4.36.0 (falls sentence-transformers Probleme macht)

# === AI & LLM ===
google-generativeai==0.3.2
openai==1.3.7  # Optional für andere LLM Provider

# === DOCUMENT PROCESSING ===
PyPDF2==3.0.1
python-docx==1.1.0
Markdown==3.5.1
beautifulsoup4==4.12.2
chardet==5.2.0

# === HTTP & UTILITIES ===
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# === LOGGING & MONITORING ===
rich==13.7.0
python-json-logger==2.0.7

# === DEVELOPMENT ===
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# === ENVIRONMENT ===
python-dotenv==1.0.0